scripts/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
# Vertex AI用（オプション）
google-cloud-aiplatform>=1.38.0